Remove spinner
